Add files via upload
In this project, "CookiesFirstDemo," you have created a dynamic web page project using Java Servlets. The project consists of two servlets, namely "FirstServlet" and "SecondServlet." These servlets handle different functionalities and interactions with the web page.

The project's front-end is designed with an HTML file called "NewHtml.html," which serves as the main web page that users interact with. This HTML file is responsible for rendering the user interface and displaying relevant information to users.

To enable interaction with a database, you have connected the project to a MySQL database using the MySQL Connector/J library. This allows you to perform database operations, such as querying and updating data, through the servlets.

Overall, "CookiesFirstDemo" is a web application that uses Java Servlets to create dynamic content and interact with users through the web page. The project's integration with a MySQL database enhances its capabilities by allowing data storage and retrieval, making it a more robust and functional web application.